requirements.txt
ftfy>=6.0.1
PyYAML>=5.4.1
argh >=0.26.2
rich>=10.9.0
numpy>=1.21.2